Add pdfverktyg fish function
The scanner that I am using cannot scan documents
that are printed on both sides. However if you scan
twice you can two PDF documents. One with the front pages
and the other with the back pages. The back pages are in
reverse order.

The pdfverktyg function allows you to call pdfverktyg.py
to create a combined pdf where the pages in the combined
pdf are in the correct order.